# kaseki-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kaseki Agent API",
    "version": "1.49.0",
    "description": "Ephemeral coding-agent runner: orchestrates Pi CLI via Docker for automated code modifications with validation and deployment",
    "contact": {
      "name": "CyanAutomation",
      "url": "https://github.com/CyanAutomation/kaseki-agent"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/CyanAutomation/kaseki-agent/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://kaseki.example.com",
      "description": "Production server (configure as needed)"
    }
  ],
  "paths": {
...
    "/api/runs/{id}/artifacts": {
      "get": {
        "operationId": "getRunArtifacts",
        "summary": "List artifacts",
        "description": "Returns metadata for all available artifacts from a run (metadata.json, git.diff, validation.log, etc.)",
        "tags": [
          "Artifacts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^kaseki-\\d+$"
            },
            "description": "Kaseki instance ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Artifacts list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "artifacts",
                    "artifactCount"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "runStatus": {
                      "type": "string",
                      "enum": [
                        "queued",
                        "running",
                        "completed",
                        "failed"
                      ]
                    },
                    "exitCode": {
                      "type": "integer"
                    },
                    "artifacts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string"
                          },
                          "size": {
                            "type": "integer"
                          },
                          "contentType": {
                            "type": "string"
                          },
                          "available": {
                            "type": "boolean"
                          },
                          "description": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "recommended": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "Recommended artifacts to review first"
                    },
                    "artifactCount": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Request ID for debugging"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Run not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Request ID for debugging"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/results/{id}/{file}": {
      "get": {
        "operationId": "downloadArtifact",
        "summary": "Download artifact file",
        "description": "Downloads a specific artifact file (e.g., git.diff, metadata.json, result-summary.md). Use `/api/runs/{id}/artifacts` to list available files.",
        "tags": [
          "Artifacts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^kaseki-\\d+$"
            },
            "description": "Kaseki instance ID"
          },
          {
            "name": "file",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Artifact filename (e.g., git.diff, metadata.json)"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Artifact file content",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Request ID for debugging"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Artifact not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Request ID for debugging"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "token",
        "description": "Bearer token authentication. Provide your API key (from KASEKI_API_KEYS environment variable) as the bearer token."
      }
    }
  },
...
}